refactor: Move ScatterGeodesicsPlotter to new file.
fixes: #178.
