+++ src/einsteinpy/tests/test_plotting/test_scattergeodesicplotter.py
@@ -28,7 +28,7 @@ def test_plot_attractor_is_called_only_once(dummy_data):
 
 
 @mock.patch(
-    "einsteinpy.plotting.geodesics_static.ScatterGeodesicPlotter._plot_attractor"
+    "einsteinpy.plotting.geodesics_scatter.ScatterGeodesicPlotter._plot_attractor"
 )
 def test_plot_calls_plot_attractor(mock_plot_attractor):
     r = [306 * u.m, np.pi / 2 * u.rad, np.pi / 2 * u.rad]
@@ -41,7 +41,7 @@ def test_plot_calls_plot_attractor(mock_plot_attractor):
     mock_plot_attractor.assert_called_with()
 
 
-@mock.patch("einsteinpy.plotting.geodesics_static.plt.show")
+@mock.patch("einsteinpy.plotting.geodesics_scatter.plt.show")
 def test_plot_show_shows_plot(mock_show):
     r = [306 * u.m, np.pi / 2 * u.rad, np.pi / 2 * u.rad]
     v = [0 * u.m / u.s, 0 * u.rad / u.s, 951.0 * u.rad / u.s]
@@ -54,7 +54,7 @@ def test_plot_show_shows_plot(mock_show):
     mock_show.assert_called_with()
 
 
-@mock.patch("einsteinpy.plotting.geodesics_static.plt.savefig")
+@mock.patch("einsteinpy.plotting.geodesics_scatter.plt.savefig")
 def test_plot_save_saves_plot(mock_save):
     r = [306 * u.m, np.pi / 2 * u.rad, np.pi / 2 * u.rad]
     v = [0 * u.m / u.s, 0 * u.rad / u.s, 951.0 * u.rad / u.s]
